refactor: IntroduceMain
