+++ src/pages/main_page/ImgMain/ImgMain.css
@@ -101,7 +101,7 @@
     /* width: 341px; */
     width: 370px;
 
-    display: inline-block, ;
+    display: inline-block;
     justify-content: left;
 }
