+++ src/pages/main_page/IntroduceMain/IntroduceMain.css
@@ -1,0 +1,77 @@
+.IntroduceMain {
+    position: relative;
+
+    font-family: 'IBMPlexSans-Regular';
+    text-align: center;
+    font-style: normal;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+
+    color: black;
+}
+
+.IntroduceMainBox {
+    position: relative;
+
+    text-align: left;
+
+    width: 1300px;
+    height: 850px;
+}
+
+
+.IntroduceMainLetterBox_1 {
+    position: absolute;
+
+    font-size: 30px;
+    font-weight: bold;
+    line-height: 150%;
+
+    top: 167px;
+}
+
+.IntroduceMainLetterBox_2 {
+    margin-top: 15px;
+
+    font-size: 20px;
+    font-weight: bold;
+}
+
+.IntroduceMainPictureBox_1 {
+    position: absolute;
+
+    width: 1300px;
+
+    justify-content: space-between;
+    display: flex;
+
+    top: 242px;
+}
+
+.IntroduceMainPictureBox_1_1 {
+    position: relative;
+
+    display: inline-block;
+}
+
+.ImgIntroduceStyle {
+    width: 390px;
+    height: 325px;
+    
+    border-radius: 10px;
+}
+
+.IntroduceMainMoreStory {
+    position: absolute;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    width: 100%;
+
+    top: 738px;
+}
