+++ src/pages/main_page/IntroduceMain/MoreStory.js
@@ -1,0 +1,16 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+import "./MoreStory.css";
+
+const MoreStory = () => {
+  return (
+    <div className="MoreStory">
+      <Link to="/introduce">
+        <button className="MoreStoryButton">더 많은 이야기 +</button>
+      </Link>
+    </div>
+  );
+};
+
+export default MoreStory;
